Export locals_without_parens for elixir formatter
This makes is so calls to the `spec_and_callback/1` macro won't be
automatically be wrapped in parentheses if the user sets up
`import_deps: [:behaves_like]` in their `.formatter.exs`. I also ran
mix format across both the main behaves_like repo and the example repo
to ensure this works. Finally, I added `mix format --check-formatted` to
both the main an example repo so you'll know if this formatting check
were to break for some reason.